feat: cleanup content.js + update manifest.json
