+++ .gitignore
@@ -1,3 +1,6 @@
+# Local
+docs/
+
 # Logs
 logs
 *.log
