+++ src/library.js
@@ -19,7 +19,7 @@ const config = {
 };
 
 // Library
-const pollPlaylistReady = (start) => {
+const pollPlaylistReady = () => {
   displayLoader();
 
   const maxPollCount = 60;
@@ -33,7 +33,7 @@ const pollPlaylistReady = (start) => {
       countUnavailableTimestamps() === countUnavailableVideos()
     ) {
       clearInterval(playlistPoll);
-      start();
+      processPlaylist();
     }
 
     pollCount++;
@@ -59,6 +59,49 @@ const displayLoader = () => {
     playlistSummary.innerHTML = "";
     playlistSummary.appendChild(loading);
   }
+};
+
+const countUnavailableTimestamps = () => {
+  const timestamps = getTimestamps(getVideos());
+  return timestamps.filter((timestamp) => timestamp === null).length;
+};
+
+const countUnavailableVideos = () => {
+  const unavailableVideoTitles = [
+    "[Private video]",
+    "[Deleted video]",
+    "[Unavailable]",
+    "[Video unavailable]",
+    "[Restricted video]",
+    "[Age restricted]"
+  ];
+
+  const videoTitles = document.querySelectorAll("a#video-title");
+
+  let unavailableVideosCount = 0;
+
+  videoTitles.forEach((videoTitle) => {
+    if (unavailableVideoTitles.includes(videoTitle.title)) {
+      unavailableVideosCount++;
+    }
+  });
+
+  return unavailableVideosCount;
+};
+
+const processPlaylist = () => {
+  configurePage();
+  setupPlaylistObserver();
+  setupEventListeners();
+  const videos = getVideos();
+  const timestamps = getTimestamps(videos);
+  const totalDurationInSeconds = timestamps.reduce((a, b) => a + b);
+  const playlistDuration = formatTimestamp(totalDurationInSeconds);
+  const playlistSummary = createPlaylistSummary({
+    timestamps,
+    playlistDuration
+  });
+  addSummaryToPage(playlistSummary);
 };
 
 const configurePage = () => {
@@ -101,34 +144,6 @@ const getVideos = () => {
   return [...videos];
 };
 
-const unformatTimestamp = (formattedTimestamp) => {
-  let timeComponents = formattedTimestamp
-    .split(":")
-    .map((timeComponent) => parseInt(timeComponent, 10));
-
-  let seconds = 0;
-  let minutes = 1;
-
-  while (timeComponents.length > 0) {
-    let timeComponent = timeComponents.pop();
-    if (isNaN(timeComponent)) continue;
-
-    seconds += minutes * timeComponent;
-    minutes *= 60;
-  }
-
-  return seconds;
-};
-
-const formatTimestamp = (timestamp) => {
-  let totalSeconds = timestamp;
-  const hours = `${Math.floor(totalSeconds / 3600)}`.padStart(2, "0");
-  totalSeconds %= 3600;
-  const minutes = `${Math.floor(totalSeconds / 60)}`.padStart(2, "0");
-  const seconds = `${totalSeconds % 60}`.padStart(2, "0");
-  return `${hours}:${minutes}:${seconds}`;
-};
-
 const getTimestamps = (videos) => {
   return videos.map((video) => {
     if (!video) return null;
@@ -144,26 +159,13 @@ const getTimestamps = (videos) => {
   });
 };
 
-const createSummaryItem = (label, value, valueColor = "#facc15") => {
-  const container = document.createElement("div");
-  container.style.margin = "8px 0px";
-  container.style.display = "flex";
-  container.style.flexDirection = "row";
-  container.style.justifyContent = "between";
-
-  const labelContainer = document.createElement("p");
-  labelContainer.textContent = label;
-
-  const valueContainer = document.createElement("p");
-  valueContainer.style.color = valueColor;
-  valueContainer.style.fontWeight = 700;
-  valueContainer.style.paddingLeft = "8px";
-  valueContainer.textContent = value;
-
-  container.appendChild(labelContainer);
-  container.appendChild(valueContainer);
-
-  return container;
+const formatTimestamp = (timestamp) => {
+  let totalSeconds = timestamp;
+  const hours = `${Math.floor(totalSeconds / 3600)}`.padStart(2, "0");
+  totalSeconds %= 3600;
+  const minutes = `${Math.floor(totalSeconds / 60)}`.padStart(2, "0");
+  const seconds = `${totalSeconds % 60}`.padStart(2, "0");
+  return `${hours}:${minutes}:${seconds}`;
 };
 
 const createPlaylistSummary = ({ timestamps, playlistDuration }) => {
@@ -242,6 +244,47 @@ const createPlaylistSummary = ({ timestamps, playlistDuration }) => {
   return summaryContainer;
 };
 
+const unformatTimestamp = (formattedTimestamp) => {
+  let timeComponents = formattedTimestamp
+    .split(":")
+    .map((timeComponent) => parseInt(timeComponent, 10));
+
+  let seconds = 0;
+  let minutes = 1;
+
+  while (timeComponents.length > 0) {
+    let timeComponent = timeComponents.pop();
+    if (isNaN(timeComponent)) continue;
+
+    seconds += minutes * timeComponent;
+    minutes *= 60;
+  }
+
+  return seconds;
+};
+
+const createSummaryItem = (label, value, valueColor = "#facc15") => {
+  const container = document.createElement("div");
+  container.style.margin = "8px 0px";
+  container.style.display = "flex";
+  container.style.flexDirection = "row";
+  container.style.justifyContent = "between";
+
+  const labelContainer = document.createElement("p");
+  labelContainer.textContent = label;
+
+  const valueContainer = document.createElement("p");
+  valueContainer.style.color = valueColor;
+  valueContainer.style.fontWeight = 700;
+  valueContainer.style.paddingLeft = "8px";
+  valueContainer.textContent = value;
+
+  container.appendChild(labelContainer);
+  container.appendChild(valueContainer);
+
+  return container;
+};
+
 const addSummaryToPage = (summary) => {
   const newDesign = isNewDesign();
 
@@ -281,34 +324,6 @@ const countTotalVideosInPlaylist = () => {
   return totalVideoCount;
 };
 
-const countUnavailableVideos = () => {
-  const unavailableVideoTitles = [
-    "[Private video]",
-    "[Deleted video]",
-    "[Unavailable]",
-    "[Video unavailable]",
-    "[Restricted video]",
-    "[Age restricted]"
-  ];
-
-  const videoTitles = document.querySelectorAll("a#video-title");
-
-  let unavailableVideosCount = 0;
-
-  videoTitles.forEach((videoTitle) => {
-    if (unavailableVideoTitles.includes(videoTitle.title)) {
-      unavailableVideosCount++;
-    }
-  });
-
-  return unavailableVideosCount;
-};
-
-const countUnavailableTimestamps = () => {
-  const timestamps = getTimestamps(getVideos());
-  return timestamps.filter((timestamp) => timestamp === null).length;
-};
-
 const isDarkMode = () => {
   return document.documentElement.getAttribute("dark") !== null;
 };
@@ -323,14 +338,4 @@ const isNewDesign = () => {
   return isNewDesign;
 };
 
-export default {
-  addSummaryToPage,
-  configurePage,
-  createPlaylistSummary,
-  formatTimestamp,
-  getTimestamps,
-  getVideos,
-  pollPlaylistReady,
-  setupEventListeners,
-  setupPlaylistObserver
-};
+export { pollPlaylistReady };
